+++ src/main/java/nextstep/subway/path/ui/PathController.java
@@ -1,7 +1,5 @@
 package nextstep.subway.path.ui;
 
-import nextstep.subway.common.exception.InvalidDataException;
-import nextstep.subway.common.exception.NoResultException;
 import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.dto.PathResponse;
 import org.springframework.http.ResponseEntity;
@@ -20,16 +18,6 @@ public class PathController {
     @GetMapping
     public ResponseEntity<PathResponse> findPaths(@RequestParam Long source, @RequestParam Long target) {
         return ResponseEntity.ok(pathService.findPath(source, target));
-    }
-
-    @ExceptionHandler(InvalidDataException.class)
-    public ResponseEntity handleInvalidDataException(InvalidDataException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(NoResultException.class)
-    public ResponseEntity handleNoResultException(NoResultException e) {
-        return ResponseEntity.badRequest().build();
     }
 
 }
